requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-environ==0.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
openai==1.3.9
feedparser==6.0.10
requests==2.31.0
httpx==0.24.0
pydantic==1.10.15
tenacity==8.2.0
asgiref==3.7.2
